Add argon2,login,register

+++ src/main/resources/static/js/usuarios.js
@@ -24,7 +24,8 @@ async function cargarUsuarios() {
 
       let botonEliminar = '<a href="#" onclick="eliminarUsuario(' + usuario.id + ')" class="btn btn-danger btn-circle btn-sm"><i class="fas fa-trash"></i></a>'
 
-      let usuarioHtml = '<tr><td>'+ usuario.id +'</td><td>'+ usuario.nombre +' '+ usuario.apellido +'</td><td>'+ usuario.email +'</td><td>'+ usuario.telefono +'</td><td>' + botonEliminar + '</td></tr>';
+       let telefono = usuario.telefono == null ? '-' : usuario.telefono;
+      let usuarioHtml = '<tr><td>'+ usuario.id +'</td><td>'+ usuario.nombre +' '+ usuario.apellido +'</td><td>'+ usuario.email +'</td><td>'+ telefono +'</td><td>' + botonEliminar + '</td></tr>';
 
 
         listadoHtml += usuarioHtml;
@@ -36,6 +37,10 @@ document.querySelector('#usuarios tbody').outerHTML = listadoHtml;
 }
 
 async function eliminarUsuario(id) {
+
+if(!confirm('¿Desea eliminar este usuario?')){
+    return;
+}
 const request = await fetch('api/usuarios/' + id, {
         method: 'DELETE',
         headers: {
@@ -44,5 +49,7 @@ const request = await fetch('api/usuarios/' + id, {
         }
 
   });
+
+  location.reload()
 
 }
